internal/errors/server.go
package errors

import (
	"errors"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
)

const defaultErrorMessage = "something went wrong"

// ServerError is used to return custom error codes to client.
type ServerError struct {
	Code    int
	Message string
	cause   error
}

func NewServerError(code int, msg string, err error) *ServerError {
	return &ServerError{
		Code:    code,
		Message: msg,
		cause:   err,
	}
}

func (s *ServerError) Error() string {
	return fmt.Sprintf("%s: %v", s.Message, s.cause)
}

func (s *ServerError) Unwrap() error {
	return s.cause
}

func GetServerErrorCode(err error) int {
	code, _, _ := ProcessServerError(err)
	return code
}

// ProcessServerError tries to retrieve from given error it's code, message and some details.
// For example, that fields can be used to build error response for client.
func ProcessServerError(err error) (code int, msg string, details string) {
	var he *echo.HTTPError
	if errors.As(err, &he) {
		return he.Code, he.Message.(string), err.Error()
	}
	var se *ServerError
	if errors.As(err, &se) {
		return se.Code, se.Message, err.Error()
	}
	return http.StatusInternalServerError, defaultErrorMessage, err.Error()
}
